+++ src/loading/loading.component.ts
@@ -1,12 +1,12 @@
-import { Component, Input, HostBinding } from '@angular/core';
+import { Component, Input, HostBinding, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
 import { inputValueToBoolean } from '../util/helpers';
 
 @Component({
     selector: 'thy-loading',
-    templateUrl: './loading.component.html'
+    templateUrl: './loading.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ThyLoadingComponent {
-
     public isDone: boolean;
 
     public tip: string;
@@ -16,20 +16,23 @@ export class ThyLoadingComponent {
     @Input()
     set thyDone(value: boolean | string) {
         this.isDone = inputValueToBoolean(value);
+        this.changeDetectorRef.markForCheck();
     }
 
     @Input()
     set thyTip(value: string) {
         this.tip = value;
+        this.changeDetectorRef.markForCheck();
     }
 
     // 不传或穿false,没有遮罩层，加载完成出现内容
     @Input()
     set thyIsMask(value: boolean | string) {
         this.isMask = inputValueToBoolean(value);
+        this.changeDetectorRef.markForCheck();
     }
 
     @HostBinding('class.thy-loading') loadingClassName = true;
 
-    constructor() { }
+    constructor(private changeDetectorRef: ChangeDetectorRef) {}
 }