+++ src/tree-select/tree-select.component.ts
@@ -17,7 +17,7 @@ import {
 } from '@angular/core';
 import { NG_VALUE_ACCESSOR, ControlValueAccessor } from '@angular/forms';
 import { ThyTreeSelectNode } from './tree-select.class';
-import { isObject, isArray } from 'util';
+import { isObject, isArray } from '../util/helpers';
 import { Observable, of } from 'rxjs';
 import {
     CdkOverlayOrigin,